+++ packages/antwerp-ui/react-components/src/lib/atoms/input/text-area/TextArea.stories.tsx
@@ -96,7 +96,7 @@ export default {
       },
       description: 'The `maxLength` attribute sets a character limitation for the value of the `TextArea`.'
     },
-    characterCountText: {
+    charCountText: {
       control: { type: 'text' },
       defaultValue: '%count% / %max%',
       table: {